+++ src/CommandHandler.php
@@ -1,9 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace BSP\CommandBus;
-
-interface CommandHandler
-{
-    public function handle($command): void;
-}