+++ tests/SimpleCommandHandlerMapTest.php
@@ -1,0 +1,26 @@
+<?php declare(strict_types=1);
+
+namespace BSP\CommandBus\Tests;
+
+use BSP\CommandBus\SimpleCommandHandlerMap;
+use BSP\CommandBus\Tests\Stub\AddSugarToCoffeeHandler;
+use PHPUnit\Framework\TestCase;
+
+final class SimpleCommandHandlerMapTest extends TestCase
+{
+    public function test handlers can be added()
+    {
+        // Given
+        $map = new SimpleCommandHandlerMap([]);
+        $this->assertSame($map->map(), []);
+        $handler = new AddSugarToCoffeeHandler();
+
+        // When
+        $map->add($handler);
+
+        // Then
+        $this->assertSame($map->map(), [
+            'BSP\CommandBus\Tests\Stub\AddSugarToCoffee' => $handler,
+        ]);
+    }
+}