+++ tests/Stub/BuyCoffee.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace BSP\CommandBus\Tests\Stub;
+
+final class BuyCoffee
+{
+}