+++ tests/Stub/BuyCoffeeHandler.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace BSP\CommandBus\Tests\Stub;
+
+final class BuyCoffeeHandler
+{
+    public function handle(BuyCoffee $command): void
+    {
+    }
+}